S&G 8
Separated game logic from command logic.
I stopped touching this project for a bit, but looking back, there's
a ton of repetitive code that only results from having to plug in
discord commands, which is kinda absurd especially when that also limits
portability to other platforms.
Extracted game logic into game.py, and plan to deprecate basically the
rest of the project other than bot.py. game.py encapsulates all logic of
the game itself, and integrations should utilize the Game class as an
instance of a lobby and utilize its instance methods to manage their
integrations.
Changing the game into a class also allows for me to potentially expand
to hosting several lobbies simultaneously in memory too, so this API is
a massive upgrade over the previous project which was kinda a day of
effort.